… best/container/object.h
… best/container/object_test.cc
… best/container/option.h
… best/container/result.h
… best/container/row.h
… best/log/internal/crash.cc
… best/math/int.h
… best/meta/empty.h
… best/text/BUILD
… best/text/format.cc
… best/text/format.h
… best/text/internal/format_impls.h
… best/text/internal/format_parser.h
… best/text/internal/utf.h
… best/text/rune.h
… best/text/str.h
… best/text/str_test.cc
… best/text/strbuf.h
… best/text/strbuf_test.cc
… best/text/utf.h
… best/text/utf16.h
… best/text/utf32.h
… best/text/utf8.h
… best/text/utf_test.cc